@@ -11,51 +11,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/cdk": "~7.3.7",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@auth0/angular-jwt": "^2.1.0",
+    "@angular/animations": "~8.0.3",
+    "@angular/cdk": "~8.0.1",
+    "@angular/common": "~8.0.3",
+    "@angular/compiler": "~8.0.3",
+    "@angular/core": "~8.0.3",
+    "@angular/forms": "~8.0.3",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.3",
+    "@angular/platform-browser-dynamic": "~8.0.3",
+    "@angular/router": "~8.0.3",
     "bootstrap": "^4.3.1",
-    "core-js": "^2.5.4",
-    "country-data": "0.0.31",
-    "countryjs": "^1.8.0",
     "hammerjs": "^2.0.8",
-    "jwt-decode": "^2.2.0",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.25",
     "ngx-cookie-service": "^2.2.0",
     "ngx-loading": "^3.0.1",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.4.0",
     "tslib": "^1.9.0",
-    "uninstall": "0.0.0",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.1",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
+    "@angular-devkit/build-angular": "~0.800.6",
+    "@angular/cli": "~8.0.6",
+    "@angular/compiler-cli": "~8.0.3",
+    "@angular/language-service": "~8.0.3",
     "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
+    "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   }
 }
