@@ -16,8 +16,8 @@ export class ClientsTableComponent implements OnInit {
   dataSource: MatTableDataSource<IclientSite>;
   clientSites = [];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(
     private apiService: ApiIntercepterService
