@@ -15,8 +15,8 @@ export class UsersTableComponent implements OnInit {
   dataSource: MatTableDataSource<IUserInfo>;
   users = [];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(private apiService: ApiIntercepterService,
     private utils: UtilsService
