@@ -18,8 +18,8 @@ export class TicketsTableComponent implements OnInit {
   dataSource: MatTableDataSource<ITicket>;
   networks: ITicket[] = [];
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(private apiService: ApiIntercepterService, private utils: UtilsService, private router: Router) {
   }
