@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { UtilsService } from './services/utils.service';
 import { timer } from 'rxjs';
+import { MatSnackBar } from '@angular/material';
 
 @Component({
   selector: 'app-root',
@@ -12,12 +13,12 @@ export class AppComponent implements OnInit {
 
   title = 'ticket-management';
   loading = false;
-  showAlert: any = false;
   loaderConfig = {
     fullScreenBackdrop: true,
   };
 
-  constructor(private router: Router, private utils: UtilsService) {
+  constructor(private router: Router, private utils: UtilsService, private snackBar: MatSnackBar) {
+
     this.utils.showLoader.subscribe(value => {
       setTimeout(() => {
         this.loading = value
@@ -31,7 +32,7 @@ export class AppComponent implements OnInit {
         const errMsg = value.data;
         let keys = Object.keys(errMsg);
         let key = keys.pop();
-        this.showAlert = `${key.replace(/\_/g, " ")}`;
+        this.snackBar.open(errMsg[key], `${key.replace(/\_/g, " ")}`, { duration: 1000 })
       }
     });
   }
